requirements.txt
requests
pandas
beautifulsoup4
ratelimit
python-dotenv
